refactor report form inputs
